restruct search
